fix DAOs e Models, primeira conexão

+++ Eleicao2026/app/src/main/java/com/example/eleicao2026/daos/VotoDAO.java
@@ -14,7 +14,7 @@ import java.util.List;
 public interface VotoDAO {
 
     @Insert
-    int inserirVoto(Voto voto); // int para retornar o id do Voto
+    long inserirVoto(Voto voto); // long para retornar o id do Voto
 
     @Insert
     void inserirRelacao(ProblemaVoto pv);
@@ -22,11 +22,11 @@ public interface VotoDAO {
     @Transaction
     default void salvarVoto(Voto v, List<Integer> codProblemas) {
 
-        int votoId = inserirVoto(v);
+        long votoId = inserirVoto(v);
 
         // para cada codigo de problema na lista
         for(Integer codPro : codProblemas) {
-            inserirRelacao(new ProblemaVoto(votoId, (int) votoId)); // insere a relação com o voto
+            inserirRelacao(new ProblemaVoto((int) votoId, (int) votoId)); // insere a relação com o voto
         }
     }
     @Query("Select Count(*) From Voto where can_codigo = :candidatoId")
